Send desired input to popmenu
