+++ .gitignore
@@ -9,6 +9,7 @@
 *.user
 *.userosscache
 *.sln.docstates
+./dist/
 
 # User-specific files (MonoDevelop/Xamarin Studio)
 *.userprefs
