+++ src/index.js
@@ -2,29 +2,22 @@ import './styles.css';
 import './images/DummyLogoTV.png';
 import movieList from './modules/variables.js';
 import renderSeries from './modules/renderNodes.js';
+import popUpInteraction from './popUp';
 
-renderSeries(movieList);
-// import { commentPopUp } from './popUp.js'
-// import { openPopUp } from './openPopUs';
-import getAPI from './apiRequest';
-import { createHTML, openPopUp } from './render';
-import closePopUP from './interactions';
+// *Don't move it!
+function openPopupWindow(cardsContainer) {
+  const commentButtons = cardsContainer.querySelectorAll('button');
+  commentButtons.forEach((button, index) => {
+    button.addEventListener('click', () => {
+      popUpInteraction(
+        movieList[index].url,
+        movieList[index].name.toLowerCase().split(' ').join('-'),
+      );
+    });
+  });
+}
 
-// const commentBtn = document.querySelector('button').addEventListener('click', commentPopUp);
-
-// const commentBtn = document.querySelector('button').addEventListener('click', openPopUp);
-const url = 'https://api.tvmaze.com/lookup/shows?imdb=tt10919420';
-
-const renderPopUp = async () => {
-  const seriesInfo = await getAPI(url);
-  const popUp = createHTML(seriesInfo);
-  const renderedHTML = openPopUp(popUp);
-  return renderedHTML;
-};
-
-const popUpInteraction = async () => {
-  const popUpWindow = await renderPopUp();
-  closePopUP(popUpWindow);
-};
-
-popUpInteraction();
+(async function main() {
+  const cardsContainer = await renderSeries(movieList);
+  openPopupWindow(cardsContainer);
+})();
